requirements.txt
arxiv
bs4
ddgs
faiss-cpu
fastapi
huggingface_hub
ipykernel
langchain
langchain-chroma
langchain-community
langchain-experimental
langchain-groq
langchain-huggingface
langchain-ollama
langchain-openai
langchain-text-splitters
langserve
mysql-connector-python
numexpr
openai
openpyxl
pymupdf
pypdf
python-dotenv
pytube
sentence_transformers
# sqlite3
sse_starlette
streamlit
uvicorn
unstructured
validators
wikipedia
youtube-transcript-api